requirements.txt
asgiref==3.4.1
boto3==1.18.27
botocore==1.21.27
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
cryptography==3.4.8
dj-database-url==0.5.0
Django==3.2.6
django-admin==2.0.1
django-cors-headers==3.7.0
django-excel-base==1.0.4
django-excel-response2==3.0.2
django-filter==2.4.0
django-heroku==0.3.1
django-redis==5.0.0
django-six==1.0.4
django-storages==1.11.1
djangorestframework==3.12.4
djangorestframework-jwt==1.9.0
djangorestframework-simplejwt==4.7.2
Faker==8.12.1
gunicorn==20.1.0
idna==3.2
jmespath==0.10.0
Markdown==3.3.4
psycopg2==2.9.1
psycopg2-binary==2.9.1
pycparser==2.20
PyJWT==1.7.1
python-dateutil==2.8.2
python-dotenv==0.19.0
pytz==2021.1
redis==3.5.3
requests==2.26.0
s3transfer==0.5.0
screen==1.0.1
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
urllib3==1.26.6
whitenoise==5.3.0
xlwt==1.3.0
